src/renderer/src/components/Song/utils.ts
import { NavigateFunction } from 'react-router-dom'
import { TSong } from '../../types'

export const fetchScreenSplit = (
  screenSplit: number | undefined,
  pageHasChords: boolean
): number => {
  return !pageHasChords ? 1 : Number(screenSplit || 6)
}

export const fetchSong = async (id: number): Promise<TSong | null> => {
  try {
    const response: TSong = await (await fetch(`http://localhost:3000/songs/${id}`)).json()
    return response
  } catch (error) {
    console.error('Error fetching song', error)
    return null
  }
}

export const goto = {
  previousPage: (
    currentPage: number,
    setCurrentPage: React.Dispatch<React.SetStateAction<number>>
  ): void => {
    setCurrentPage(currentPage - 1)
  },

  nextPage: (
    currentPage: number,
    setCurrentPage: React.Dispatch<React.SetStateAction<number>>
  ): void => {
    setCurrentPage(currentPage + 1)
  },

  titlePage: (setCurrentPage: React.Dispatch<React.SetStateAction<number>>): void => {
    setCurrentPage(0)
  },

  previousSong: (currentSong: number, Navigate: NavigateFunction): void => {
    Navigate(`/song/${Math.max(currentSong - 1, 0)}`)
  },

  nextSong: (currentSong: number, Navigate: NavigateFunction): void => {
    Navigate(`/song/${currentSong + 1}`)
  },

  repertoire: (Navigate: NavigateFunction): void => {
    Navigate('/repertoire')
  },

  setList: (Navigate: NavigateFunction): void => {
    Navigate('/setlist')
  },

  gigList: (Navigate: NavigateFunction): void => {
    Navigate('/')
  },

  toggleTimerFreeze: (
    timerHalted: boolean,
    setTimerHalted: React.Dispatch<React.SetStateAction<boolean>>
  ): void => {
    setTimerHalted(!timerHalted)
  }
}
